MAINT: use 32-bit ints to store coordinates
[CHANGED] I'm assuming no single coordinate type has an position
    > 2^31

+++ src/ensembl_cli/convert.py
@@ -8,7 +8,8 @@ from cogent3.core.location import LostSpan, Map, Span
 def seq_to_gap_coords(seq: Sequence) -> numpy.ndarray:
     """returns coordinates of sequence gaps"""
     m, x = seq.parse_out_gaps()
-    return numpy.array(m.get_gap_coordinates(), dtype=int)
+    # Assuming the maximum integer is < 2^31
+    return numpy.array(m.get_gap_coordinates(), dtype=numpy.int32)
 
 
 def gap_coords_to_seq(coords: numpy.ndarray, ungapped: Sequence) -> Aligned:
